Update scripting.md
Added build-essentials to Linux
Moved dependencies above quick start
Updated grammar
Dropped inappropriate "you"

When one does college level writing the use of the word "you" when referring to the reader of a general publication is considered inappropriate. Replacing with the word "you" with "one" or restructuring the sentence is more appropriate.